PHOENIX-7481 HBase 3 compatibility changes: Cleanup deprecated APIs, HTable and HTableDescriptor
