+++ phoenix-core-client/src/main/java/org/apache/phoenix/execute/DelegateHTable.java
@@ -24,7 +24,6 @@ import java.util.concurrent.TimeUnit;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hbase.CompareOperator;
-import org.apache.hadoop.hbase.HTableDescriptor;
 import org.apache.hadoop.hbase.TableName;
 import org.apache.hadoop.hbase.client.Append;
 import org.apache.hadoop.hbase.client.CheckAndMutate;
@@ -44,7 +43,6 @@ import org.apache.hadoop.hbase.client.Table;
 import org.apache.hadoop.hbase.client.TableDescriptor;
 import org.apache.hadoop.hbase.client.coprocessor.Batch.Call;
 import org.apache.hadoop.hbase.client.coprocessor.Batch.Callback;
-import org.apache.hadoop.hbase.filter.CompareFilter.CompareOp;
 import org.apache.hadoop.hbase.ipc.CoprocessorRpcChannel;
 import org.apache.phoenix.compat.hbase.CompatDelegateHTable;
 
@@ -70,18 +68,8 @@ public class DelegateHTable extends CompatDelegateHTable implements Table {
     }
 
     @Override
-    public HTableDescriptor getTableDescriptor() throws IOException {
-        return delegate.getTableDescriptor();
-    }
-
-    @Override
     public boolean exists(Get get) throws IOException {
         return delegate.exists(get);
-    }
-
-    @Override
-    public boolean[] existsAll(List<Get> gets) throws IOException {
-        return delegate.existsAll(gets);
     }
 
     @Override
@@ -132,18 +120,6 @@ public class DelegateHTable extends CompatDelegateHTable implements Table {
     }
 
     @Override
-    public boolean checkAndPut(byte[] row, byte[] family, byte[] qualifier, byte[] value, Put put)
-            throws IOException {
-        return delegate.checkAndPut(row, family, qualifier, value, put);
-    }
-
-    @Override
-    public boolean checkAndPut(byte[] row, byte[] family, byte[] qualifier, CompareOp compareOp,
-            byte[] value, Put put) throws IOException {
-        return delegate.checkAndPut(row, family, qualifier, compareOp, value, put);
-    }
-
-    @Override
     public void delete(Delete delete) throws IOException {
         delegate.delete(delete);
     }
@@ -151,18 +127,6 @@ public class DelegateHTable extends CompatDelegateHTable implements Table {
     @Override
     public void delete(List<Delete> deletes) throws IOException {
         delegate.delete(deletes);
-    }
-
-    @Override
-    public boolean checkAndDelete(byte[] row, byte[] family, byte[] qualifier, byte[] value,
-            Delete delete) throws IOException {
-        return delegate.checkAndDelete(row, family, qualifier, value, delete);
-    }
-
-    @Override
-    public boolean checkAndDelete(byte[] row, byte[] family, byte[] qualifier, CompareOp compareOp,
-            byte[] value, Delete delete) throws IOException {
-        return delegate.checkAndDelete(row, family, qualifier, compareOp, value, delete);
     }
 
     @Override
@@ -231,77 +195,8 @@ public class DelegateHTable extends CompatDelegateHTable implements Table {
     }
 
     @Override
-    public boolean checkAndMutate(byte[] row, byte[] family, byte[] qualifier, CompareOp compareOp,
-            byte[] value, RowMutations mutation) throws IOException {
-        return delegate.checkAndMutate(row, family, qualifier, compareOp, value, mutation);
-    }
-
-    @Override
-    public void setOperationTimeout(int operationTimeout) {
-        delegate.setOperationTimeout(operationTimeout);
-    }
-
-    @Override
-    public int getOperationTimeout() {
-        return delegate.getOperationTimeout();
-    }
-
-    @Override
-    public int getRpcTimeout() {
-        return delegate.getRpcTimeout();
-    }
-
-    @Override
-    public void setRpcTimeout(int rpcTimeout) {
-        delegate.setRpcTimeout(rpcTimeout);
-    }
-
-    @Override
     public TableDescriptor getDescriptor() throws IOException {
         return delegate.getDescriptor();
-    }
-
-    @Override
-    public boolean checkAndPut(byte[] row, byte[] family, byte[] qualifier, CompareOperator op,
-            byte[] value, Put put) throws IOException {
-        return delegate.checkAndPut(row, family, qualifier, op, value, put);
-    }
-
-    @Override
-    public boolean checkAndDelete(byte[] row, byte[] family, byte[] qualifier, CompareOperator op,
-            byte[] value, Delete delete) throws IOException {
-        return delegate.checkAndDelete(row, family, qualifier, op, value, delete);
-    }
-
-    @Override
-    public CheckAndMutateBuilder checkAndMutate(byte[] row, byte[] family) {
-        return delegate.checkAndMutate(row, family);
-    }
-
-    @Override
-    public boolean checkAndMutate(byte[] row, byte[] family, byte[] qualifier, CompareOperator op,
-            byte[] value, RowMutations mutation) throws IOException {
-        return delegate.checkAndMutate(row, family, qualifier, op, value, mutation);
-    }
-
-    @Override
-    public int getReadRpcTimeout() {
-        return delegate.getReadRpcTimeout();
-    }
-
-    @Override
-    public void setReadRpcTimeout(int readRpcTimeout) {
-        delegate.setReadRpcTimeout(readRpcTimeout);
-    }
-
-    @Override
-    public int getWriteRpcTimeout() {
-        return delegate.getWriteRpcTimeout();
-    }
-
-    @Override
-    public void setWriteRpcTimeout(int writeRpcTimeout) {
-        delegate.setWriteRpcTimeout(writeRpcTimeout);
     }
 
     @Override
