+++ phoenix-core/src/it/java/org/apache/phoenix/end2end/index/BaseIndexIT.java
@@ -48,7 +48,6 @@ import org.apache.hadoop.hbase.Cell;
 import org.apache.hadoop.hbase.CellScanner;
 import org.apache.hadoop.hbase.CellUtil;
 import org.apache.hadoop.hbase.HConstants;
-import org.apache.hadoop.hbase.client.HTable;
 import org.apache.hadoop.hbase.client.Result;
 import org.apache.hadoop.hbase.client.ResultScanner;
 import org.apache.hadoop.hbase.client.Scan;
@@ -914,7 +913,7 @@ public abstract class BaseIndexIT extends ParallelStatsDisabledIT {
             conn.commit();
 	            
             // the index table is one row
-            HTable table = (HTable) conn.unwrap(PhoenixConnection.class).getQueryServices().getTable(fullTableName.getBytes());
+            Table table = conn.unwrap(PhoenixConnection.class).getQueryServices().getTable(fullTableName.getBytes());
             ResultScanner resultScanner = table.getScanner(new Scan());
             for (Result result : resultScanner) {
             	System.out.println(result);
