+++ phoenix-hbase-compat-2.4.1/src/main/java/org/apache/phoenix/compat/hbase/CompatOmidTransactionTable.java
@@ -18,13 +18,11 @@ package org.apache.phoenix.compat.hbase;
 
 import java.io.IOException;
 
-import org.apache.hadoop.hbase.HTableDescriptor;
 import org.apache.hadoop.hbase.client.Delete;
 import org.apache.hadoop.hbase.client.Put;
 import org.apache.hadoop.hbase.client.Result;
 import org.apache.hadoop.hbase.client.RowMutations;
 import org.apache.hadoop.hbase.client.Table;
-import org.apache.hadoop.hbase.client.Table.CheckAndMutateBuilder;
 import org.apache.hadoop.hbase.filter.CompareFilter.CompareOp;
 
 public abstract class CompatOmidTransactionTable implements Table {
@@ -33,11 +31,6 @@ public abstract class CompatOmidTransactionTable implements Table {
 
     public CompatOmidTransactionTable(Table hTable) {
         this.hTable = hTable;
-    }
-
-    @Override
-    public HTableDescriptor getTableDescriptor() throws IOException {
-        return hTable.getTableDescriptor();
     }
 
     @Override
